HPCC-19584 Add support for new field filters into JHTree
Refactor SegMonitorList to put it behind an interface, and provide an
alternative implementation using new field filters.

Remove unused code from SegMonitorList and SegMonitors.

Moved test code out of rtlnewkey.cpp as it was preventing indexer from working
in Eclipse.

Default createValueSets to 1 in compiler.

Clean up IIndexReadContext interface to remove methods that should not have
been there.

Reduce memory footprint of IndexFilterList by using a RowInfo designed for
fixed offsets only.

Update serializeTypes due to more optimal serialization of ifblock condition.
