+++ common/thorhelper/thorxmlwrite.cpp
@@ -121,137 +121,6 @@ void CommonFieldProcessor::processEndRow(const RtlFieldInfo * field)
 
 
 //=============================================================================================
-// MORE - this function should probably move into IIndexReadContext interface rather than leaking ordinality() and item() out of that interface just for me
-void printKeyedValues(StringBuffer &out, IIndexReadContext *segs, IOutputMetaData *rowMeta)
-{
-    unsigned totalKeyedSize = 0;
-    unsigned numSegs = segs->ordinality();
-    while (numSegs)
-    {
-        IKeySegmentMonitor &seg = *segs->item(numSegs-1);
-        if (!seg.isWild())
-        {
-            totalKeyedSize = seg.getOffset() + seg.getSize();
-            break;
-        }
-        numSegs--;
-    }
-    if (numSegs)
-    {
-        byte *tempRow = (byte *) alloca(totalKeyedSize);
-        byte *savedRow = (byte *) alloca(totalKeyedSize);
-        const RtlFieldInfo * const *fields = rowMeta->queryTypeInfo()->queryFields();
-        unsigned fieldOffset = 0;
-        bool inKeyed = false;
-        bool inWild = false;
-        for (unsigned segNo = 0; segNo < numSegs; segNo++)
-        {
-            IKeySegmentMonitor &seg = *segs->item(segNo);
-            unsigned segOffset = seg.getOffset();
-            unsigned segSize = seg.getSize();
-            while (fieldOffset < segOffset + segSize) // This is trying to cope with the combined case but not sure it completely does
-            {
-                assertex(fields[0]->type->isFixedSize());
-                unsigned curFieldSize = fields[0]->type->size(NULL, NULL);
-                if (seg.isWild())
-                {
-                    if (!inWild)
-                    {
-                        if (inKeyed)
-                        {
-                            out.append("),");
-                            inKeyed = false;
-                        }
-                        out.append("WILD(");
-                        inWild = true;
-                    }
-                    else
-                        out.append(',');
-                    out.append(fields[0]->name);
-                }
-                else
-                {
-                    StringBuffer setValues;
-                    CommonFieldProcessor setProcessor(setValues, true);
-                    unsigned numValues = 0;
-                    unsigned subStringLength = 0;
-                    if (!seg.isEmpty())
-                    {
-                        seg.setLow(tempRow);
-                        for (;;)
-                        {
-                            if (numValues)
-                                setValues.append(",");
-                            memcpy(savedRow+segOffset, tempRow+segOffset, segSize);
-                            seg.endRange(tempRow);
-                            if (memcmp(savedRow+segOffset, tempRow+segOffset, segSize) != 0)
-                            {
-                                // Special case - if they differ only in trailing values that are 0 vs 0xff, then it's a substring match...
-                                if (numValues==0 && (fields[0]->type->fieldType & (RFTMkind | RFTMebcdic)) == type_string)
-                                {
-                                    unsigned pos;
-                                    for (pos = 0; pos < segSize; pos++)
-                                    {
-                                        if (savedRow[segOffset+pos] != tempRow[segOffset+pos])
-                                            break;
-                                    }
-                                    subStringLength = pos;
-                                    for (; pos < segSize; pos++)
-                                    {
-                                        if (savedRow[segOffset+pos] != 0 || tempRow[segOffset+pos] != 0xff)
-                                        {
-                                            subStringLength = 0;
-                                            break;
-                                        }
-                                    }
-                                }
-                                fields[0]->process(savedRow + fieldOffset, tempRow, setProcessor);
-                                setValues.append("..");
-                                fields[0]->process(tempRow + fieldOffset, tempRow, setProcessor);
-                                numValues+=2;
-                            }
-                            else
-                            {
-                                fields[0]->process(tempRow + fieldOffset, tempRow, setProcessor);
-                                numValues++;
-                            }
-                            if (!seg.increment(tempRow))
-                                break;
-                        }
-                    }
-                    if (!inKeyed)
-                    {
-                        if (inWild)
-                        {
-                            out.append("),");
-                            inWild = false;
-                        }
-                        out.append("KEYED(");
-                        inKeyed = true;
-                    }
-                    else
-                        out.append(',');
-                    out.append(fields[0]->name);
-                    if (numValues==1)
-                        out.append("=").append(setValues);
-                    else if (subStringLength)
-                        out.appendf("[1..%d]='", subStringLength).append(subStringLength, (char *) savedRow+fieldOffset).append("'");
-                    else
-                        out.append(" IN [").append(setValues).append("]");
-                }
-                fieldOffset += curFieldSize;
-                fields++;
-                if (!fields[0])
-                    break;
-            }
-        }
-        if (inKeyed || inWild)
-            out.append(")");
-    }
-    else
-        out.append("UNKEYED");
-}
-
 
 extern thorhelper_decl void convertRowToXML(size32_t & lenResult, char * & result, IOutputMetaData & info, const void * row, unsigned flags)
 {
