+++ system/jhtree/bloom.hpp
@@ -113,7 +113,7 @@ public:
      */
     IndexBloomFilter(unsigned numHashes, unsigned tableSize, byte *table, __uint64 fields);
     inline __int64 queryFields() const { return fields; }
-    bool reject(const SegMonitorList &segs) const;
+    bool reject(const IIndexFilterList &filters) const;
     static int compare(CInterface *const *a, CInterface *const *b);
 private:
     const __uint64 fields;
@@ -160,7 +160,6 @@ extern jhtree_decl IBloomBuilder *createBloomBuilder(const IBloomBuilderInfo &_h
 interface IRowHasher : public IInterface
 {
     virtual hash64_t hash(const byte *row) const = 0;
-    virtual bool isExact(const SegMonitorList &segs) const = 0;
     virtual __uint64 queryFields() const = 0;
 };
 
@@ -175,11 +174,11 @@ extern jhtree_decl IRowHasher * createRowHasher(const RtlRecord &recInfo, __uint
 /**
  * Retrieve bloom/partition hash corresponding to a supplied filter condition
  * @param fields   Bitmap containing field numbers
- * @param segs     Filter to be checked
+ * @param filter   Filter to be checked
  * @param hash     Initial hash value, updated to reflect supplied fields
  * return          true if the filter is suitable for bloom filtering/partitioning via returned hash value
  */
-extern jhtree_decl bool getBloomHash(__int64 fields, const SegMonitorList &segs, hash64_t &hashval);
+extern jhtree_decl bool getBloomHash(__int64 fields, const IIndexFilterList &filter, hash64_t &hashval);
 
 
 #endif
