+++ testing/regress/ecl/key/serializetypes.xml
@@ -950,10 +950,10 @@
  <Row><Result_9>true</Result_9></Row>
 </Dataset>
 <Dataset name='Result 10'>
- <Row><Result_10>507AF1CE43427384F94E0100000101000008000000000104040000002B0440000001003D020601030101666F72656E616D650002040400002B0440000000013D048C01000000000000000004000000000000000091017375726E616D650002040400000D0440000904686173666F72656E616D650001000000007375726E616D65636F756E7400000101000000032B04000000042B0400001402100000052B0440000000003D0A060100000000000000030100000000000000060300000000000000030300000000000000060500000000000000030500000000000000060700000000000000030700000000000000060C00000000000000030C000000000000000165787472610002040400000D04400026056964000001010000666972737400050D0400007365636F6E6400050D0400006E65737465647265636F72640006140280006E65737465647265636F726401526F770000072B040000</Result_10></Row>
+ <Row><Result_10>50D925F894F3BFCC004B0100000101000008000000000104040000002B0440000001003D010101666F72656E616D650002040400002B0440000000013D048C01000000000000000004000000000000000091017375726E616D650002040400000D0440000904686173666F72656E616D650001000000007375726E616D65636F756E7400000101000000032B04000000042B0400001402100000052B0440000000003D0A060100000000000000030100000000000000060300000000000000030300000000000000060500000000000000030500000000000000060700000000000000030700000000000000060C00000000000000030C000000000000000165787472610002040400000D04400026056964000001010000666972737400050D0400007365636F6E6400050D0400006E65737465647265636F72640006140280006E65737465647265636F726401526F770000072B040000</Result_10></Row>
 </Dataset>
 <Dataset name='Result 11'>
- <Row><Result_11>507AF1CE43427384F94E0100000101000008000000000104040000002B0440000001003D020601030101666F72656E616D650002040400002B0440000000013D048C01000000000000000004000000000000000091017375726E616D650002040400000D0440000904686173666F72656E616D650001000000007375726E616D65636F756E7400000101000000032B04000000042B0400001402100000052B0440000000003D0A060100000000000000030100000000000000060300000000000000030300000000000000060500000000000000030500000000000000060700000000000000030700000000000000060C00000000000000030C000000000000000165787472610002040400000D04400026056964000001010000666972737400050D0400007365636F6E6400050D0400006E65737465647265636F72640006140280006E65737465647265636F726401526F770000072B040000</Result_11></Row>
+ <Row><Result_11>50D925F894F3BFCC004B0100000101000008000000000104040000002B0440000001003D010101666F72656E616D650002040400002B0440000000013D048C01000000000000000004000000000000000091017375726E616D650002040400000D0440000904686173666F72656E616D650001000000007375726E616D65636F756E7400000101000000032B04000000042B0400001402100000052B0440000000003D0A060100000000000000030100000000000000060300000000000000030300000000000000060500000000000000030500000000000000060700000000000000030700000000000000060C00000000000000030C000000000000000165787472610002040400000D04400026056964000001010000666972737400050D0400007365636F6E6400050D0400006E65737465647265636F72640006140280006E65737465647265636F726401526F770000072B040000</Result_11></Row>
 </Dataset>
 <Dataset name='Result 12'>
  <Row><Result_12>true</Result_12></Row>
