+++ tools/dumpkey/dumpkey.cpp
@@ -194,7 +194,7 @@ int main(int argc, const char **argv)
                 {
                     writer.setown(new SimpleOutputWriter);
                     const RtlRecord &inrec = diskmeta->queryRecordAccessor(true);
-                    manager.setown(createLocalKeyManager(inrec, index, nullptr));
+                    manager.setown(createLocalKeyManager(inrec, index, nullptr, false));
                     size32_t minRecSize = 0;
                     if (globals->hasProp("fields"))
                     {
